# Modify
- Glide 처리 수정
- Dagger : 2.12 -> 2.13
- build:gradle : 3.0.1

+++ app/src/main/java/com/pluu/webtoon/MyGlideModule.kt
@@ -19,7 +19,7 @@ class MyGlideModule : AppGlideModule() {
     private val DISK_CACHE_SIZE = 1024 * 1024 * 10
 
     override fun applyOptions(context: Context, builder: GlideBuilder) {
-        builder.setMemoryCache(LruResourceCache(cacheSize))
+        builder.setMemoryCache(LruResourceCache(cacheSize.toLong()))
                 .setDiskCache(ExternalCacheDiskCacheFactory(context, "cache", DISK_CACHE_SIZE))
     }
